Refactor Android manifest into internal/templates
This is just a simple cleanup that moves the template to
the common templates package and also moves the data struct to
where it is being used.
