+++ cmd/fyne/internal/templates/bundled.go
@@ -5,6 +5,11 @@ package templates
 
 import "fyne.io/fyne/v2"
 
+var resourceAndroidManifestXml = &fyne.StaticResource{
+	StaticName: "AndroidManifest.xml",
+	StaticContent: []byte(
+		"<manifest\n\txmlns:android=\"http://schemas.android.com/apk/res/android\"\n\tpackage=\"{{.JavaPkgPath}}\"\n\tandroid:versionCode=\"{{.Build}}\"\n\tandroid:versionName=\"{{.Version}}\">\n\n\t<application android:label=\"{{.Name}}\" android:debuggable=\"{{.Debug}}\">\n\t<activity android:name=\"org.golang.app.GoNativeActivity\"\n\t\tandroid:label=\"{{.Name}}\"\n\t\tandroid:configChanges=\"orientation|keyboardHidden|uiMode\"\n\t\tandroid:theme=\"@android:style/Theme\">\n\t\t<meta-data android:name=\"android.app.lib_name\" android:value=\"{{.LibName}}\" />\n\t\t<intent-filter>\n\t\t\t<action android:name=\"android.intent.action.MAIN\" />\n\t\t\t<category android:name=\"android.intent.category.LAUNCHER\" />\n\t\t</intent-filter>\n\t</activity>\n\t</application>\n\n\t<uses-permission android:name=\"android.permission.WRITE_EXTERNAL_STORAGE\" />\n\t<uses-permission android:name=\"android.permission.READ_EXTERNAL_STORAGE\" />\n\t<uses-permission android:name=\"android.permission.INTERNET\" />\n</manifest>"),
+}
 var resourceInfoPlist = &fyne.StaticResource{
 	StaticName: "Info.plist",
 	StaticContent: []byte(
