+++ cmd/fyne/internal/templates/templates.go
@@ -20,6 +20,9 @@ var (
 	// FyneMetadataInit is the metadata injecting file for fyne metadata
 	FyneMetadataInit = template.Must(template.New("fyne_metadata_init.got").Parse(string(resourceFynemetadatainitGot.StaticContent)))
 
+	// ManifestAndroid is the default manifest for building an Android application
+	ManifestAndroid = template.Must(template.New("AndroidManifest").Parse(string(resourceAndroidManifestXml.StaticContent)))
+
 	// ManifestWindows is the manifest file for windows packaging
 	ManifestWindows = template.Must(template.New("Manifest").Parse(string(resourceAppManifest.StaticContent)))
